fix: upgrade scripts

+++ scripts/localnode.sh
@@ -15,11 +15,18 @@ TRACE=""
 HOME_DIR=mytestnet
 DENOM=ppica
 
-# remove existing daemon
-rm -rf $HOME_DIR
+
+if [ "$CONTINUE" == "true" ]; then
+    echo "\n ->> continuing from previous state"
+    $BINARY start --home $HOME_DIR --log_level debug
+    exit 0
+fi
 
 $BINARY config keyring-backend $KEYRING
 $BINARY config chain-id $CHAINID
+
+# remove existing daemon
+rm -rf $HOME_DIR
 
 # if $KEY exists it should be deleted
 echo "decorate bright ozone fork gallery riot bus exhaust worth way bone indoor calm squirrel merry zero scheme cotton until shop any excess stage laundry" | $BINARY keys add $KEY --keyring-backend $KEYRING --algo $KEYALGO --recover --home $HOME_DIR
